+++ elements/user-picker/src/__tests__/unit/components/InputSpec.tsx
@@ -13,12 +13,29 @@ describe('ClearIndicator', () => {
     expect(component.prop('isDisabled')).toBeFalsy();
   });
 
-  it('should be disabled if disableInput select prop is passed in', () => {
-    const component = shallowInput({
-      selectProps: { disableInput: true },
+  it('should fire event.preventDefault() if isDisabled', () => {
+    const noop = () => {};
+    const mockedProps = {
+      innerRef: (ref: React.Ref<HTMLInputElement>) => {},
+      getStyles: noop,
+      cx: noop,
+      selectProps: {
+        disableInput: true,
+      },
+    };
+    const spiedPreventDefault = jest.fn();
+    const component = mount(<Input {...mockedProps} />);
+    component.find('input').simulate('keyPress', {
+      key: 'a',
+      preventDefault: spiedPreventDefault,
     });
-
-    expect(component.prop('isDisabled')).toBeTruthy();
+    expect(spiedPreventDefault).toHaveBeenCalledTimes(1);
+    // Backspace shoud still register for deleting selected users
+    component.find('input').simulate('keyPress', {
+      key: 'Backspace',
+      preventDefault: spiedPreventDefault,
+    });
+    expect(spiedPreventDefault).toHaveBeenCalledTimes(2);
   });
 
   it('should fire event.preventDefault() only on `Enter` key pressed', () => {