+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 34.4.9
+
+### Patch Changes
+
+- Updated dependencies
+
+## 34.4.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 34.4.7
 
 ### Patch Changes